refactor(server): agent logic and modules
