+++ packages/server/src/modules/app.module.ts
@@ -5,7 +5,7 @@ import { OpenAIModule } from './openai/openai.module';
 import { DatabaseModule } from './databases/database.module';
 import { MessageModule } from './message/message.module';
 import { GrowlyModule } from './growly/growly.module';
-import { ProxyMiddleware } from './middleware/x402-redirect.middleware';
+import { ProxyMiddleware } from '../middleware/x402-redirect.middleware';
 import { SuiteCoreModule } from './suite-core/suite-core.module';
 
 @Module({
