+++ packages/server/src/modules/chat/chat.module.ts
@@ -1,5 +1,5 @@
 import { Module } from '@nestjs/common';
-import { MessageModule } from 'src/message/message.module';
+import { MessageModule } from '../message/message.module';
 import { AgentModule } from '../agent/agent.module';
 import { ChatController } from './chat.controller';
 import { ChatService } from './chat.service';
