+++ packages/server/src/modules/message/message.interface.ts
