+++ packages/server/src/modules/message/message.module.ts
@@ -1,8 +1,8 @@
 import { Module } from '@nestjs/common';
 import { MessageRepository } from './message.repository';
 import { MessageService } from './message.service';
-import { DatabaseModule } from 'src/databases/database.module';
-import { OpenAIModule } from 'src/openai/openai.module';
+import { DatabaseModule } from '../databases/database.module';
+import { OpenAIModule } from '../openai/openai.module';
 
 @Module({
   imports: [DatabaseModule, OpenAIModule],
