+++ packages/server/src/modules/suite-core/suite-core.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { SuiteCoreController } from './suite-core.controller';
 import { SuiteCoreService } from './suite-core.service';
 import { ConfigModule } from '@nestjs/config';
-import { DatabaseModule } from 'src/databases/database.module';
+import { DatabaseModule } from '../databases/database.module';
 
 @Module({
   imports: [ConfigModule, DatabaseModule],
